Update superchain safe
