+++ apps/dapp-console/app/layout.tsx
@@ -1,4 +1,3 @@
-import Head from 'next/head'
 import { Inter } from 'next/font/google'
 import '@/app/globals.css'
 import { PrivyProviderWrapper } from '@/app/providers/PrivyProviderWrapper'
@@ -13,26 +12,8 @@ export default function RootLayout({
 }: Readonly<{
   children: React.ReactNode
 }>) {
-  const metadata = {
-    title: 'Superchain Dev Console',
-    description:
-      'Tools to help you build, launch, and grow your dapp on the Superchain',
-    image: '/banners/page-banner.png',
-  }
-
   return (
     <html lang="en">
-      <Head>
-        <title>{metadata.title}</title>
-        <meta name="description" content={metadata.description} />
-        <meta property="og:title" content={metadata.title} />
-        <meta property="og:description" content={metadata.description} />
-        <meta property="og:image" content={metadata.image} />
-        <meta name="twitter:title" content={metadata.title} />
-        <meta name="twitter:description" content={metadata.description} />
-        <meta name="twitter:image" content={metadata.image} />
-        <meta name="twitter:card" content="summary_large_image" />
-      </Head>
       <body className={inter.className}>
         <FeatureFlagProvider>
           <ThemeProvider attribute="class" defaultTheme="system">
